+++ feature-teacher/approve-outing/src/commonMain/kotlin/com.b1nd.dodam.approve-outing/module/ViewModelModule.kt
@@ -1,5 +1,6 @@
-package com.b1nd.dodam.approve
+package com.b1nd.dodam.approveouting
 
+import com.b1nd.dodam.approveouting.viewmodel.ApproveOutViewModel
 import org.koin.compose.viewmodel.dsl.viewModel
 import org.koin.dsl.module
