+++ feature-teacher/approve-outing/src/commonMain/kotlin/com.b1nd.dodam.approve-outing/viewmodel/ApproveOutViewModel.kt
@@ -1,7 +1,10 @@
-package com.b1nd.dodam.approve
+package com.b1nd.dodam.approveouting.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.b1nd.dodam.approve.ApproveOutState
+import com.b1nd.dodam.approve.DetailMember
+import com.b1nd.dodam.approve.OutPendingUiState
 import com.b1nd.dodam.common.date.DodamDate
 import com.b1nd.dodam.common.result.Result
 import com.b1nd.dodam.common.utiles.combineWhenAllComplete
